[FIX] 권한 없을때 무한루프 해결 및 컴포넌트 분리
