+++ package.json
@@ -26,7 +26,9 @@
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
-    "json-server": "json-server --watch db.json --port 5000"
+    "json-server": "json-server --watch db.json --port 5000",
+    "deploy": "yarn run build && aws s3 sync build/ s3://balance-game-bucket"
+
   },
   "eslintConfig": {
     "extends": [
